knowledge_api/requirements.txt
fastapi>=0.68.0
pydantic>=1.8.0
uvicorn>=0.15.0
llama_index>=0.9.7
tiktoken>=0.4.0
psycopg2-binary
openai>=1.3.5
python-dotenv>=1.0.0
python-multipart>=0.0.6
pypdf>=3.16.2
pymilvus>=2.3.1